remove tool execution assertion
